Add a diagram showing where the barrier feet are
After some thought, I believe we should be able to construct the
barrier with only 8 feet (i.e. 2 per side).  These are positioned such
that there is an equal mass of barrier to both sides of them.

When we come to construct this barrier, we can place some support of
the barrier as follows:

 1) Diagonal bracing just on-top of the barrier from the top of one
    leg to the nearest one on the nearest side.

 2) The corners of the barrier can be rigidly attached by vertical
    supports to the scaff diagonals that the arena features.
    Furthermore, diagonal struts can be placed from the top of this
    vertical support down to the tops of the two nearest legs.

Hopefully, this will mean that we can construct this barrier such that
the 3m segments are quickly removable.  This'll make it considerably
easier to walk around the arena.