Remove erroneous marker-blob lines from markers in docs
There was a bug in the marker generation script that resulted in a
line being drawn from the end of the marker's text to the principle
corner blob.  This line is invisible on the screen, but comes out in
printing.  This patch removes these lines from the markers in the
rules.