State the pre-match procedure more fully
I've created these rules based on our previous experience. The 'staging
area' works really well to ensure everyone knows what's going on and
who's about to enter the arena. It also provides a perfect place for the
robot badges to be applied and checked.

I moved the '1 minute before the match' rule into this block, as it
seems fitting. I also reduced this to 40 seconds due to the higher rate
of matches we will have at the competition this year.