Reduce the severity of the match forfeit if slow rule
It seems a bit pointless having the 40 seconds if it's never possible to
use it. Obviously it would only ever be allowed for very, very quick
fixes and we can still kick them out of the arena if we can see it's
gonna take too long. Basically that's what we've always done, so we
might as well make the rules match reality a bit more.