Move up the rule about tokens starting upside down.
This introduces the idea of tokens have a right-way-up, which is
otherwise a surprise to the reader while they're trying to understand
the scoring rules (part d mentions orientation without otherwise
explaining that tokens have such a concept).
It also moves a comment about the start of the match before the game
points rules, which are about the end of a match.

Change-Id: Id7bafa32dcd659b42e48a2f0ac4776e95e77283c