Clarify placement of a token in a robot before the match.
Remove the conflicting mention of this from the specs section, as
the version of it there was incorrect (vs what was elsewhere in the
rules, and vs what had been agreed on the mailing list), and redundant.
Tidy the wording of the mention of this in the main rules section,
to ensure that they can only place one of their own tokens, and
that it has to be from one of the near_est_ rows.

I believe that all these changes are within the original intent of
the rules.

Change-Id: I176e779f32a10fee3a68cb46dada746b92ad469c