Fix #2088: Clarify token colours and placement.
This updates the images of the arena to use colours similar to those
we've been using on the badges, and to clarify which tokens go where.
Modify the item about where the tokens go to explain the placement
scheme, and then use this scheme to update the colours in the diagrams.

2014-02-27 Patch set 2: rebased
2014-02-27 Patch set 3: Simplify description in favour of just pointing
                        at the diagram. Explicitly mention the innermost
                        positions so it's possible to see  which
                        colour belongs to which corner.
2014-02-28 Patch set 4: rebased
2014-03-13 Patch set 5: Replace incorrect references to 'zones' with
                        'corners' instead.

Change-Id: I71b79b7d85e402a0dccdc7ac83a4e25b720cda41