Order the arena markers in a clockwise fashion.
The arena corners have always (as far as I recall) been ordered
clockwise, and the outer markers anti-clockwise. Particularly when
adding other counted elements to the arena (such as the slots in
SR2014), this creates confusion about which way around things are
ordered.

Change the marker ordering to match the corners, standardising the
default arena to being clockwise. This consistency both ensures
clarity and ease of remembering which way around things go.

I've chosen clockwise as some of the website's SRComp pages render
the team's corner positions in the 4-box shape of the arena, and
keeping those the same avoids needing to change that. I'm not aware
of anything else which renders the arena, nor would need to change
to match this.

This patch deliberately doesn't bother changing the slots since
they're not relevant for future games anyway.

Change-Id: I4383a13c72e957a4eecdd1d5fbfdd13ad94112b0