Loosen up arena wall specification
Our existing arena design is somewhat incompatible with the number of
spectators that we expect in our knock-out matches.  Solutions to this
problem involve moving the audience back, and onto tiered seating, etc.

In order to make it possible to be seated/stood back from the arena and
still see robots that are at its edge, the arena wall will need to be
shorter.  Also, it seems highly likely that it won't be made out of white
hardboard.

So this patch changes the wall height constraint to be 100mm minimum,
rather than specified as 600mm.  This allows us the flexibility to adapt
the arena's construction to the situation it's in.  (e.g. it may be that
arena "A" has a higher audience capacity than arena "B", and so the
latter has a 'standard' hardboard wall, etc.)

This sounds like it leaves the arena markers hovering in free space.  Of
course, these can be supported by any structure of our choosing.

This creates the possibility that robots will snag on the edges of the
arena markers.  We could set the arena markers back from the surface of
the wall by 50mm to reduce this problem, and potentially move them up by
50mm to clear the minimum wall height.

Change-Id: I30e6847727755f6ba6abec78cec56d0ff915f270