Increase the number of token markers
The markers do not identify which zone a token belongs to, they only
serve to give each token a unique ID. The major constraint is that these
IDs *are* unique: we cannot have to two tokens with the same markers in
the arena.

This makes spares more difficult, because if a token breaks we have to
replace it with a token with the same marker number. If we want to build
spares ahead-of-time, we have to build spares per token position, rather
than having a single pool of spares.

By increasing the number of token IDs available, we can have one token
with each ID at the competition. In case of breakages, we can just swap
in any of the spares.