Fix #2631: Make the gap between zones bigger
By making the zone gap (and barrier width) 400 rather than 300mm,
we ensure that even a diagonally placed flag could not possibly be
in contact with more than one zone at a time. In turn this makes
the current state of the game easier to see and faster to judge.

Change-Id: I5dcc525d3447cf2a2d2ee1856848639b8edd1cbd