Resolve TODO by moving footnote away from summary
The general summary of the game doesn't need to include the detail
of what makes a token as 'up', though clearly this is needed when
describing how points are allocated. Rather than do battle with
LaTeX over trying to get the footnote in both places, just put it
in the latter.

Patch set 2: don't even use footnote.

Change-Id: Ifa0eddac4ef9d961c67cbe3ebddde4e96cb0e17a