Specify the token nets arrangements
Re-label them with letters to help disambiguate them from the
various other things we're labelling. (This also further suggests
to a reader that their properties are not related to the other
numbered things). Add a diagram adapted from Ed's ones, but
adjusted to our more usual colours and combinging all three nets
into a single diagram.

Change-Id: I95719249a259d0497cd845640a13a6f9a80f8ba9