Add thumbnail make target
This adds a target which creates a target suitable for use in srweb
on the rules page. This doesn't need updating every time the rules
are (and wasn't actually what I've used in the past to create it)
but is useful to have as an automated command rather than needing
to manually do this via GIMP etc.

Change-Id: I299c9e307bbb48852f42df843b65b1a685189961