Update safety regs regarding battery cable
The "large green power connector" is a feature of the V3 power board,
which is no longer used. Since the introduction of the V4 power board
in SR2015, the battery cable is anchored to the power board by
ring terminals sealed inside the power board's case.

It was deemed that checking the amount of exposed copper at the
boundary between the cable and the large green power connector is
no longer a relevant test, since the cables should be covered by
insulation up to and through the wall of the power board's case.
Any damage to this insulation would be covered by the following test.