Change the size of the markers and smaller tokens to make them fit
Previously the actual size of the markers was 125mm (due to the
necessary white border), which doesn't fit on the currently specified
tokens. Resizing both ensures that the specification actually works.

The tokens we're looking at getting are 127mm internal side length
boxes, which can be up to 137mm outer length in some cases.